Refactor: Separate header file for class
